requirements.txt
aiohttp
beautifulsoup4
pyppeteer
Pillow
apng
onnxruntime